darknet_yolov4/requirements.txt
Pillow==7.1.0
seldon-core==1.4.0
